requirements.txt
flask 
tensorflow
numpy
gunicorn
Pillow
Flask-Cors